Add bin/find-last, a script to find the most recent version of a file with a glob name.
You might be like, wtf are you thinking. Well, if instead of making a
library or something, you just keep copying and pasting a file into the
repo, but then tweaking it to make it better, you might have a bunch
versions of the file with the same name hanging around, but you
want to know which version is good.

If you happen to be in this exact situation, DO I HAVE THE SCRIPT FOR YOU